Move OntologyTreeSearchContainer.tsx out of OntologyBrowserPanel.tsx
